move dummydata file, edit dummydata code to reflect modularized interface
